feat: Introduce centralized authentication error handling and refactor related imports
